Update Client4 tests
